LayoutTests/fast/dom/collection-children-backwards.html

<!DOCTYPE html>
<html>
<body>
<div id="container">
    <div id="child0"></div>
    <div id="child1"></div>
    <div id="child2"></div>
</div>
<script src="../../resources/js-test.js"></script>
<script>

description('This tests iterating element.children backwarwds. The collection index getter should return the last element when indexed for the last item.');
shouldBe('container.children[3]; container.children[2]', 'child2');

</script>
</body>
</html>
